requirements.txt
absl-py==0.11.0
APScheduler==3.6.3
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageai==2.1.6
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==7.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.4.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
